refactor: move logic to auth wrapper
