+++ src/App.tsx
@@ -1,37 +1,25 @@
 import { Routes } from "react-router-dom";
 import { RoutesList } from "./config/routes";
 import Layout from "./app/Layout";
-import { ReactNode, useState } from "react";
+import { ReactNode } from "react";
 import { QueryClient, QueryClientProvider } from "react-query";
 import { ReactQueryDevtools } from "react-query/devtools";
 import { env } from "./config/env";
 import { Toaster } from "./components/ui/sonner";
-import { useEffect } from "react";
-import { useRefreshToken } from "./features/authentication/hooks/use-refresh-token";
+import { AuthWrapper } from "./features/authentication/wrapper";
 
 const queryClient = new QueryClient();
 
 export default function App(): ReactNode {
-  const [isRefreshing, setIsRefreshing] = useState<boolean>(false);
-  const refreshToken = useRefreshToken();
-
-  useEffect(() => {
-    if (!isRefreshing) {
-      setIsRefreshing(true);
-      refreshToken().finally(() => {
-        setIsRefreshing(false);
-      });
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [refreshToken]);
-
   return (
     <QueryClientProvider client={queryClient}>
-      <Layout>
-        <Routes>{RoutesList}</Routes>
-        {env.app.mode === "development" && <ReactQueryDevtools />}
-        <Toaster />
-      </Layout>
+      <AuthWrapper>
+        <Layout>
+          <Routes>{RoutesList}</Routes>
+          {env.app.mode === "development" && <ReactQueryDevtools />}
+          <Toaster />
+        </Layout>
+      </AuthWrapper>
     </QueryClientProvider>
   );
 }
