+++ src/features/authentication/wrapper/index.ts
@@ -1,0 +1,1 @@
+export { AuthWrapper } from "./AuthWrapper";
